Extract yaml file uploading into its own component.
